version: Add comments and add error code return
Add comments to the version.c file in the function early_hostname()
and add error code return to the function early_hostname() to indicate if
the hostname has been truncated or not.

If it returns -EINVAL then the hostname has been truncated and the user
should be notified of this by a pr_warn() message. If it returns 0 then
the hostname has not been truncated and the user isnt notified of this.

Signed-off-by: lewisevans2007 <lewisevans2007@users.noreply.github.com>