initramfs: Add pr_warn for failure
The function free_hash() in initramfs.c is responsible for freeing the
initramfs hash table. However, in certain scenarios, the function may
not be able to free all the data in the hash table as expected.

To alert the user about this situation, we have added a pr_warn
statement inside the free_hash() function.

This warning message will be printed in case the hash table still
contains data after the function completes its execution, which might
indicate a potential issue with memory management in the initramfs
module. By adding this warning, this can help provide valuable
information to users and developers to investigate and resolve
any memory-related  problems related to the hash table in the
initramfs code.

Signed-off-by: lewisevans2007 <lewisevans2007@users.noreply.github.com>