printk_clear_debug: Add a macro
This commit introduces the printk_clear_debug macro, which facilitates the printing of kernel debugging messages in a clear and easily identifiable format in the log. This macro is designed to aid debugging in situations with a high volume of log messages, making it simpler to identify relevant debugging output.

The printk_clear_debug() macro includes special formatting to enclose the debug message within "[ DEBUG ]" markers, making it more prominent in the log. Additionally, the macro can display messages in the log even when the debug level is turned off, providing valuable debugging information regardless of the current log level settings.

Furthermore, the macro is configurable based on the kernel configuration option CONFIG_PRINTK_CLEAR_DEBUG. If this option is enabled, the printk_clear_debug macro will be active and display the debugging messages. On the other hand, if the option is disabled, the macro will be defined as an empty statement, preventing the messages from showing up in the log.

This change enables developers to selectively control the visibility of the debugging messages by configuring the CONFIG_PRINTK_CLEAR_DEBUG option during the kernel build process.

Example:

printk_clear_debug("test") // No need for a new line

Output:

[    0.000000] ====================[ DEBUG ]====================
[    0.000000] test
[    0.000000] ==================[ END DEBUG ]==================

Signed-off-by: lewisevans2007 <lewisevans2007@users.noreply.github.com>